Refactoring and using Rapido in prompts controller
