+++ app/controllers/api/prompts_controller.rb
@@ -1,46 +1,29 @@
 require 'prompt'
 
 class Api::PromptsController < ActionController::API
-  before_action :load_workflow
+  include Rapido::Controller
+  include Rapido::ApiController
 
-  rescue_from Workflow::UnmatchableDataError do |_e|
-    render json: { error: 'No matching step found.', facts: fact_params.to_h }, status: 422
+  rescue_from Prompt::Factory::UnmatchableDataError do |e|
+    render json: { error: e }, status: 422
   end
 
-  rescue_from Prompt::Parts::AnswerNotFound do |e|
+  rescue_from Prompt::PartsCollection::AnswerNotFound do |e|
     render json: { error: e }, status: 500
   end
 
-  rescue_from Prompt::Answers::AnswerNotFound do |e|
+  rescue_from Prompt::AnswersCollection::AnswerNotFound do |e|
     render json: { error: e }, status: 500
   end
 
-  def create
-    Rails.logger.info "New step request for workflow #{params[:workflow_id]}: #{fact_params}"
-    step = @workflow.match(fact_params)
-    prompt = Prompt.new(step)
-    render json: prompt.to_h(fact_params.to_h)
+  rescue_from Rapido::Errors::RecordNotFound do |e|
+    render json: { error: 'Workflow not found.' }, status: 404
   end
 
-  private
+  owner_class :workflow
+  owner_lookup_param :workflow_token
+  owner_lookup_field :token
 
-  def load_workflow
-    @workflow = Workflow.find_by(token: params[:workflow_token])
-    return if @workflow
-    render json: { error: 'Workflow not found' }, status: :not_found
-  end
-
-  def fact_params
-    params[:facts].present? ? params.require(:facts).permit! : {}
-  end
-
-  def step
-    @step = {}
-    @step[:text] = 'My first name is [:user_first_name], and my last name is [:user_last_name].'
-    @step[:answers] = {
-      user_first_name: { input_type: 'short_text' },
-      user_last_name: { input_type: 'short_text' }
-    }
-    @step
-  end
+  free_from_authority!
+  permit_all_params!
 end
