+++ features/api/workflows/auth.feature
@@ -23,7 +23,7 @@ Feature: Simulated authentication and introduction flow
       ]
     }
     """
-    And the response status should be "200"
+    And the response status should be "201"
 
   Scenario:
     When I send a POST request to "/api/workflows/auth/prompts" with the following:
@@ -34,7 +34,7 @@ Feature: Simulated authentication and introduction flow
       }
     }
     """
-    Then the response status should be "200"
+    Then the response status should be "201"
     And the JSON response should be:
     """
     {
@@ -58,7 +58,7 @@ Feature: Simulated authentication and introduction flow
       }
     }
     """
-    Then the response status should be "200"
+    Then the response status should be "201"
     And the JSON response should be:
     """
     {
