+++ features/api/workflows/jsonpath.feature
@@ -11,7 +11,7 @@ Feature: JSONPath
       "facts": {}
     }
     """
-    Then the response status should be "200"
+    Then the response status should be "201"
     And the JSON response should be:
     """
     {
