+++ features/support/env.rb
@@ -3,7 +3,7 @@ require 'rspec/matchers'
 require 'spreewald/all_steps'
 require 'cucumber/api_steps'
 
-ActionController::Base.allow_rescue = false
+ActionController::Base.allow_rescue = true
 ActiveRecord::Migration.maintain_test_schema!
 begin
   DatabaseCleaner.strategy = :transaction
